api/requirements.txt
alembic==1.12.0
blinker==1.6.2
click==8.1.7
Flask==2.3.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mysql-connector-python==8.1.0
packaging==23.1
protobuf==4.21.12
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.7
